Make all links open in the same frame, to make it consistent.
It used to be a convention, that links outside the page would open in a new tab/window. Maybe we can make all the links in the footer like that.

However, I think most of the folks (especially those using Testacular) know how to use CMD+click to open a link in a new tab. And therefore there's no reason for target="_blank". Let the users control what they want.